refactor: reorg of core for cleaner import strat
