+++ jaclang/core/construct.py
@@ -2,14 +2,15 @@
 
 from __future__ import annotations
 
+import shelve
 import unittest
+from contextvars import ContextVar
 from dataclasses import dataclass, field
 from typing import Callable, Optional
 from uuid import UUID, uuid4
 
 from jaclang.compiler.constant import EdgeDir
 from jaclang.core.utils import collect_node_connections
-from jaclang.plugin.feature import JacFeature as Jac
 from jaclang.plugin.spec import DSFunc
 
 
@@ -57,6 +58,8 @@ class NodeAnchor(ObjectAnchor):
 
     def populate_edges(self) -> None:
         """Populate edges from edge ids."""
+        from jaclang.plugin.feature import JacFeature as Jac
+
         if len(self.edges) == 0 and len(self.edge_ids) > 0:
             for e_id in self.edge_ids:
                 edge = Jac.context().get_obj(e_id)
@@ -354,11 +357,15 @@ class NodeArchitype(Architype):
 
     def __init__(self) -> None:
         """Create node architype."""
+        from jaclang.plugin.feature import JacFeature as Jac
+
         self._jac_: NodeAnchor = NodeAnchor(obj=self)
         Jac.context().save_obj(self, persistent=self._jac_.persistent)
 
     def save(self) -> None:
         """Save the node to the memory/storage hierarchy."""
+        from jaclang.plugin.feature import JacFeature as Jac
+
         self._jac_.persistent = True
         Jac.context().save_obj(self, persistent=True)
 
@@ -383,11 +390,15 @@ class EdgeArchitype(Architype):
 
     def __init__(self) -> None:
         """Create edge architype."""
+        from jaclang.plugin.feature import JacFeature as Jac
+
         self._jac_: EdgeAnchor = EdgeAnchor(obj=self)
         Jac.context().save_obj(self, persistent=self.persistent)
 
     def save(self) -> None:
         """Save the edge to the memory/storage hierarchy."""
+        from jaclang.plugin.feature import JacFeature as Jac
+
         self.persistent = True
         Jac.context().save_obj(self, persistent=True)
 
@@ -405,6 +416,8 @@ class EdgeArchitype(Architype):
 
     def populate_nodes(self) -> None:
         """Populate nodes for the edges from node ids."""
+        from jaclang.plugin.feature import JacFeature as Jac
+
         if self._jac_.source_id:
             obj = Jac.context().get_obj(self._jac_.source_id)
             if obj is None:
@@ -439,6 +452,13 @@ class WalkerArchitype(Architype):
         self._jac_: WalkerAnchor = WalkerAnchor(obj=self)
 
 
+class GenericEdge(EdgeArchitype):
+    """Generic Root Node."""
+
+    _jac_entry_funcs_ = []
+    _jac_exit_funcs_ = []
+
+
 class Root(NodeArchitype):
     """Generic Root Node."""
 
@@ -460,11 +480,157 @@ class Root(NodeArchitype):
         self._jac_.edges = []
 
 
-class GenericEdge(EdgeArchitype):
-    """Generic Root Node."""
+class Memory:
+    """Memory module interface."""
 
-    _jac_entry_funcs_ = []
-    _jac_exit_funcs_ = []
+    mem: dict[UUID, Architype] = {}
+    save_obj_list: dict[UUID, Architype] = {}
+
+    def __init__(self) -> None:
+        """init."""
+        pass
+
+    def get_obj(self, obj_id: UUID) -> Architype | None:
+        """Get object from memory."""
+        return self.get_obj_from_store(obj_id)
+
+    def get_obj_from_store(self, obj_id: UUID) -> Architype | None:
+        """Get object from the underlying store."""
+        ret = self.mem.get(obj_id)
+        return ret
+
+    def has_obj(self, obj_id: UUID) -> bool:
+        """Check if the object exists."""
+        return self.has_obj_in_store(obj_id)
+
+    def has_obj_in_store(self, obj_id: UUID) -> bool:
+        """Check if the object exists in the underlying store."""
+        return obj_id in self.mem
+
+    def save_obj(self, item: Architype, persistent: bool) -> None:
+        """Save object."""
+        self.mem[item._jac_.id] = item
+        if persistent:
+            # TODO: check if it needs to be saved, i.e. dirty or not
+            self.save_obj_list[item._jac_.id] = item
+
+    def commit(self) -> None:
+        """Commit changes to persistent storage, if applicable."""
+        pass
+
+    def close(self) -> None:
+        """Close any connection, if applicable."""
+        self.mem.clear()
+
+
+class ShelveStorage(Memory):
+    """Shelve storage for jaclang runtime object."""
+
+    storage: shelve.Shelf | None = None
+
+    def __init__(self, session: str = "") -> None:
+        """Init shelve storage."""
+        super().__init__()
+        if session:
+            self.connect(session)
+
+    def get_obj_from_store(self, obj_id: UUID) -> Architype | None:
+        """Get object from the underlying store."""
+        obj = super().get_obj_from_store(obj_id)
+        if obj is None and self.storage:
+            obj = self.storage.get(str(obj_id))
+            if obj is not None:
+                self.mem[obj_id] = obj
+
+        return obj
+
+    def has_obj_in_store(self, obj_id: UUID | str) -> bool:
+        """Check if the object exists in the underlying store."""
+        return obj_id in self.mem or (
+            str(obj_id) in self.storage if self.storage else False
+        )
+
+    def commit(self) -> None:
+        """Commit changes to persistent storage."""
+        if self.storage is not None:
+            for obj_id, obj in self.save_obj_list.items():
+                self.storage[str(obj_id)] = obj
+        self.save_obj_list.clear()
+
+    def connect(self, session: str) -> None:
+        """Connect to storage."""
+        self.session = session
+        self.storage = shelve.open(session)
+
+    def close(self) -> None:
+        """Close the storage."""
+        super().close()
+        self.commit()
+        if self.storage:
+            self.storage.close()
+        self.storage = None
+
+
+class ExecutionContext:
+    """Default Execution Context implementation."""
+
+    mem: Optional[Memory]
+    root: Optional[Root]
+
+    def __init__(self) -> None:
+        """Create execution context."""
+        super().__init__()
+        self.mem = ShelveStorage()
+        self.root = None
+
+    def init_memory(self, session: str = "") -> None:
+        """Initialize memory."""
+        if session:
+            self.mem = ShelveStorage(session)
+        else:
+            self.mem = Memory()
+
+    def get_root(self) -> Root:
+        """Get the root object."""
+        if self.mem is None:
+            raise ValueError("Memory not initialized")
+
+        if not self.root:
+            root = self.mem.get_obj(UUID(int=0))
+            if root is None:
+                self.root = Root()
+                self.mem.save_obj(self.root, persistent=self.root._jac_.persistent)
+            elif not isinstance(root, Root):
+                raise ValueError(f"Invalid root object: {root}")
+            else:
+                self.root = root
+        return self.root
+
+    def get_obj(self, obj_id: UUID) -> Architype | None:
+        """Get object from memory."""
+        if self.mem is None:
+            raise ValueError("Memory not initialized")
+
+        return self.mem.get_obj(obj_id)
+
+    def save_obj(self, item: Architype, persistent: bool) -> None:
+        """Save object to memory."""
+        if self.mem is None:
+            raise ValueError("Memory not initialized")
+
+        self.mem.save_obj(item, persistent)
+
+    def reset(self) -> None:
+        """Reset the execution context."""
+        if self.mem:
+            self.mem.close()
+        self.mem = None
+        self.root = None
+
+
+exec_context: ContextVar[ExecutionContext | None] = ContextVar(
+    "ExecutionContext", default=None
+)
 
 
 class JacTestResult(unittest.TextTestResult):
